+++ lib/src/db/isoexpress/rt_energy_offer.dart
@@ -2,17 +2,16 @@ library db.isoexpress.rt_energy_offer;
 
 import 'dart:convert';
 import 'dart:io';
+import 'package:archive/archive.dart';
 import 'package:csv/csv.dart';
 import 'package:duckdb_dart/duckdb_dart.dart';
-import 'package:elec/elec.dart';
 import 'package:elec_server/client/isoexpress/energy_offer.dart';
 import 'package:logging/logging.dart';
+import 'package:more/collection.dart';
 import 'package:path/path.dart';
-import 'package:timezone/timezone.dart';
 import 'package:date/date.dart';
 import '../lib_iso_express.dart';
-import '../converters.dart';
-import 'package:elec_server/src/utils/iso_timestamp.dart';
+import 'package:dotenv/dotenv.dart' as dotenv;
 
 class RtEnergyOfferArchive {
   RtEnergyOfferArchive({required this.dir});
@@ -27,112 +26,37 @@ class RtEnergyOfferArchive {
   String getUrl(Date asOfDate) =>
       'https://webservices.iso-ne.com/api/v1.1/hbrealtimeenergyoffer/day/${yyyymmdd(asOfDate)}';
 
-  File getFilename(Date asOfDate) =>
-      File('$dir/Raw/hbrealtimeenergyoffer_${yyyymmdd(asOfDate)}.json');
-
-  /// [rows] has the data for all the hours of the day for one asset
-  Map<String, dynamic> converter(List<Map> rows) {
-    var row = <String, dynamic>{};
-
-    /// daily info
-    row['date'] = formatDate(rows.first['Day']);
-    row['Masked Lead Participant ID'] =
-        rows.first['Masked Lead Participant ID'];
-    row['Masked Asset ID'] = rows.first['Masked Asset ID'];
-    row['Must Take Energy'] = rows.first['Must Take Energy'];
-    row['Maximum Daily Energy Available'] =
-        rows.first['Maximum Daily Energy Available'];
-    row['Unit Status'] = rows.first['Unit Status'];
-    row['Claim 10'] = rows.first['Claim 10'];
-    row['Claim 30'] = rows.first['Claim 30'];
-
-    /// hourly info
-    row['hours'] = [];
-    for (var hour in rows) {
-      var aux = <String, dynamic>{};
-      var utc = parseHourEndingStamp(hour['Day'], hour['Trading Interval']);
-      aux['hourBeginning'] = TZDateTime.fromMicrosecondsSinceEpoch(
-              IsoNewEngland.location, utc.microsecondsSinceEpoch)
-          .toIso8601String();
-      aux['Economic Maximum'] = hour['Economic Maximum'];
-      aux['Economic Minimum'] = hour['Economic Minimum'];
-      aux['Cold Startup Price'] = hour['Cold Startup Price'];
-      aux['Intermediate Startup Price'] = hour['Intermediate Startup Price'];
-      aux['Hot Startup Price'] = hour['Hot Startup Price'];
-      aux['No Load Price'] = hour['No Load Price'];
-
-      /// add the non empty price/quantity pairs
-      var pricesHour = <num?>[];
-      var quantitiesHour = <num?>[];
-      for (var i = 1; i <= 10; i++) {
-        if (hour['Segment $i Price'] is! num) break;
-        pricesHour.add(hour['Segment $i Price']);
-        quantitiesHour.add(hour['Segment $i MW']);
-      }
-      aux['price'] = pricesHour;
-      aux['quantity'] = quantitiesHour;
-      row['hours'].add(aux);
-    }
-    return row;
-  }
+  File getFilename(Date asOfDate) => File(
+      '$dir/Raw/${asOfDate.year}/hbrealtimeenergyoffer_${yyyymmdd(asOfDate)}.json.gz');
 
   List<EnergyOfferSegment> processFile(File file,
       {String extension = '.json'}) {
-    if (extension == '.csv') {
-      return _processFileCsv(file);
-    } else if (extension == '.json') {
-      return _processFileJson(file);
-    } else {
-      throw ArgumentError('File type $extension not supported');
-    }
+    return _processFileJson(file);
   }
 
   List<EnergyOfferSegment> _processFileJson(File file) {
-    final aux = file.readAsStringSync();
-    final data = json.decode(aux) as Map<String, dynamic>;
+    final bytes = file.readAsBytesSync();
+    var content = GZipDecoder().decodeBytes(bytes);
+    var aux = utf8.decoder.convert(content);
+    var data = json.decode(aux) as Map<String, dynamic>;
+
     var out = <EnergyOfferSegment>[];
     if (data['HbRealTimeEnergyOffers'] == '') return out;
 
-    final offers =
-        data['HbRealTimeEnergyOffers']['HbRealTimeEnergyOffer'] as List;
-    for (Map<String, dynamic> offer in offers) {
-      out.addAll(EnergyOfferSegment.fromJson(offer));
+    late List<Map<String, dynamic>> xs;
+
+    final offers = data['HbRealTimeEnergyOffers'];
+    if (offers is Map) {
+      xs = (data['HbRealTimeEnergyOffers']['HbRealTimeEnergyOffer'] as List)
+          .cast<Map<String, dynamic>>();
+    } else {
+      throw StateError('Unsupported format');
     }
 
-    return out;
-  }
-
-  List<EnergyOfferSegment> _processFileCsv(File file) {
-    // var data = mis.readReportTabAsMap(file, tab: 0);
-    // if (data.isEmpty) return [];
-    // var dataByAssetId = groupBy(data, (dynamic row) => row['Masked Asset ID']);
-    // var out = dataByAssetId.keys
-    //     .map((ptid) => converter(dataByAssetId[ptid]!))
-    //     .toList();
-    return <EnergyOfferSegment>[];
-  }
-
-  /// Aggregate all the days of the month in
-  ///
-  List<EnergyOfferSegment> aggregateDays(List<Date> days) {
-    var out = <EnergyOfferSegment>[];
-    for (var date in days) {
-      log.info('...  Working on $date');
-      final file = getFilename(date);
-      if (file.existsSync()) {
-        var rows = file
-            .readAsLinesSync()
-            .map((e) => const CsvToListConverter().convert(e).first);
-        var offers = rows
-            .where((row) => row.first == 'D')
-            .expand((row) => EnergyOfferSegment.fromRow(row));
-        out.addAll(offers);
-        log.info('...  Added ${rows.length} rows');
-      } else {
-        throw StateError('Missing file for $date');
-      }
+    for (Map<String, dynamic> x in xs) {
+      out.addAll(EnergyOfferSegment.fromJson(x));
     }
-    log.info('aggregated data has ${out.length} rows!');
+
     return out;
   }
 
@@ -196,7 +120,7 @@ CREATE TABLE IF NOT EXISTS rt_offers (
     UnitStatus ENUM('ECONOMIC', 'UNAVAILABLE', 'MUST_RUN') NOT NULL,
 );  
   ''');
-      for (var month in months) {
+    for (var month in months) {
       // remove the data if it's already there
       con.execute('''
 DELETE FROM rt_offers 
@@ -216,5 +140,33 @@ FROM read_csv(
     con.close();
 
     return 0;
+  }
+
+  Future<void> downloadDay(Date day) async {
+    var user = dotenv.env['ISONE_WS_USER']!;
+    var pwd = dotenv.env['ISONE_WS_PASSWORD']!;
+
+    var client = HttpClient()
+      ..addCredentials(
+          Uri.parse(getUrl(day)), '', HttpClientBasicCredentials(user, pwd))
+      ..userAgent = 'Mozilla/4.0'
+      ..badCertificateCallback = (cert, host, port) {
+        print('Bad certificate connecting to $host:$port:');
+        return true;
+      };
+    var request = await client.getUrl(Uri.parse(getUrl(day)));
+    request.headers.set(HttpHeaders.acceptHeader, 'application/json');
+    var response = await request.close();
+    var fileName = getFilename(day).path.removeSuffix('.gz');
+    if (!Directory(dirname(fileName)).existsSync()) {
+      Directory(dirname(fileName)).createSync(recursive: true);
+    }
+    await response.pipe(File(fileName).openWrite());
+    // gzip it
+    var res = Process.runSync('gzip', ['-f', fileName], workingDirectory: dir);
+    if (res.exitCode != 0) {
+      throw StateError('Gzipping ${basename(fileName)} has failed');
+    }
+    log.info('Downloaded and gzipped file ${basename(fileName)}');
   }
 }