+++ test/db/lib_update_dbs_test.dart
@@ -52,11 +52,12 @@ Future<void> tests() async {
   // await updateIsoneRtSystemLoad5minArchive(days: days, download: true);
 
   // await insertDays(DaLmpHourlyArchive(), days, gzip: true);
-  final months = Month(2022, 1, location: IsoNewEngland.location)
-      .upTo(Month(2024, 9, location: IsoNewEngland.location));
+  final months = Month(2024, 7, location: IsoNewEngland.location)
+      .upTo(Month(2024, 7, location: IsoNewEngland.location));
   // await updateIsoneDaLmp(months: months, download: true);
-  await updateIsoneDemandBids(months: months, download: false);
   // await updateIsoneMonthlyAssetNcpc(months: months, download: true);
+  // await updateIsoneDemandBids(months: months, download: false);
+  await updateIsoneRtEnergyOffers(months: months, download: false);
 
   // await updateIsoneRtLmp(months: months, download: true);
   // await updateIsoneRtLmp5Min(